test: Add extra memory test
Cleanup cloud fixture as it's not needed.
Add test for checking memory retained after interactions
